Update item edit and delete design
Removed pencil icon for editing a checklist item and updated item edit logic to maintain consistency in design between editing a checklist title and editing a checklist item. Also updated trash can icon for deleting an item to a clear icon to improve overall UI consistency of a checklist card.